Convert application to newer versions of Backbone and Underscore
In the process, change pattern used to initialize view modules
so that individually, they are each loaded after the DOM
element that will contain them has been rendered. This allows
the views to set thier 'el' properties correctly when they
are initialized, bypassing the convoluted method used to
set 'el' on the views previously.