Fix problem with synonym list horizonal positioning
Introduced by use of a variable-width font. Now a constant
character width value cannot be used. Instead, the y position
for a synonym list is determined by copying the text before the
root word (to the start of line) into a hidden div which has the
same font/style/size and taking that div's width. This way the
browser can calculate the correct width, and place the synonym
list accurately.